Small bug fix for constants in on-the-fly message classes (#33)
Prior to this commit constant values like "1.0", i.e., floats that are
integer, were not correctly treated in the md5sum calculation -- they
were abbreviated to "1", causing incorrect md5sums. This commit fixes
that by using the original string value for md5sum calculation (up to
the start of a comment (for strings) or the first whitespace character
for any other field type).

- also fixing some wrong logic for null values