Fix bug with hasParam and getParamNames.
The function "methodCall" was being incorrectly called on an object of
type XmlrpcClient. "methodCall" is actually defined for a property of
instances of XmlrpcClient, so this was a simple issue of missing indirection.

XmlrpcClient is a wrapper that defines a "call" function as an interface, and
delegates to an object from the xmlrpc module by calling its "methodCall".
This commit changes the implementatino of hasParam and getParamNames to
use "call" instead of "methodCall", adding the missing layer of indirection.