Fix backticks in message definitions, throw better error on missing dependencies. (#205)
* Message gen: escape backticks in comments

Until now, backticks ("`") in comments in message definitions will lead to an exception when the generated message is loaded, because the generated JS code is syntactically invalid.

* Message generation: indicate name of missing dependencies

Previously, if a custom message definition referred to another package that was missing, the error was not giving any clues as to where the problem occurred, i.e., which packages are involved. This change fixes that, making it much easy to debug broken dependencies.